move tests into src/__tests__
