+++ src/schemas/__tests__/Array.test.js
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-import { normalize, schema } from '../../src';
+import { normalize, schema } from '../../';
 
 describe(schema.Array.name, () => {
   it('normalizes an array of objects', () => {
