+++ src/schemas/__tests__/Object.test.js
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-import { normalize, schema } from '../../src';
+import { normalize, schema } from '../../';
 
 describe(schema.Object.name, () => {
   it('normalizes an object', () => {
